+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 0.2
 * Use `setTimeout` instead of `setInterval` to avoid speeding-up issues
+* Added option to use custom slides instead of plain images
+* Added options to include titles and captions with plain images
+* Added option to only show first slide on mobile version
 
 ## 0.1
 * First version