+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changelog for Pilau Slideshow
 
+## 0.3
+* Added ability to choose between navigation arrow linking options
+
 ## 0.2
 * Use `setTimeout` instead of `setInterval` to avoid speeding-up issues
 * Added option to use custom slides instead of plain images